Extract url construction into a new class
